♻️ Refactor ts to nodenext module
