+++ .eslintrc
@@ -28,6 +28,7 @@
         "location": "start"
       }
     ],
-    "@typescript-eslint/explicit-module-boundary-types": "off"
+    "@typescript-eslint/explicit-module-boundary-types": "off",
+    "node/no-missing-import": "off"
   }
 }
