+++ packages/cli/src/flags/errors.ts
@@ -1,4 +1,3 @@
-// eslint-disable-next-line node/no-missing-import
 import { Errors } from '@oclif/core';
 
 export class IncorrectFlagValue extends Errors.CLIError {}
