+++ packages/cli/src/flags/id.ts
@@ -1,8 +1,6 @@
-// eslint-disable-next-line node/no-missing-import
 import { Flags, type Interfaces } from '@oclif/core';
 import { getConfig } from '@lokse/core';
-import { MissingFlagValue } from './errors';
-import type { Action } from './errors';
+import { MissingFlagValue, type Action } from './errors.js';
 
 export const flag = Flags.string({
     char: 'i',
