+++ packages/cli/src/index.ts
@@ -1,4 +1,3 @@
-// eslint-disable-next-line node/no-missing-import
 export { run } from '@oclif/core';
 
 export type { LokseConfig } from '@lokse/core';
