+++ packages/core/src/index.ts
@@ -1,20 +1,20 @@
-export { Reader, WorksheetReader, InvalidFilterError } from './reader';
+export { Reader, WorksheetReader, InvalidFilterError } from './reader/index.js';
 
-export type { WorksheetLinesByTitle, SheetsFilter } from './reader';
+export type { WorksheetLinesByTitle, SheetsFilter } from './reader/index.js';
 
-export type { Logger } from './logger';
+export type { Logger } from './logger.js';
 
-export { transformersByFormat } from './transformer';
+export { transformersByFormat } from './transformer/index.js';
 
-export { default as FileWriter } from './writer';
-export { default as Line } from './line';
+export { default as FileWriter } from './writer.js';
+export { default as Line } from './line.js';
 
-export { OutputFormat, NAME } from './constants';
+export { OutputFormat, NAME } from './constants.js';
 
-export { FatalError } from './errors';
+export { FatalError } from './errors.js';
 
-export * from './config';
+export * from './config.js';
 
-export * from './plugins';
+export * from './plugins/index.js';
 
-export * from './utils';
+export * from './utils.js';
