+++ packages/core/src/reader/__tests__/worksheet-reader.test.ts
@@ -1,5 +1,5 @@
 import { GoogleSpreadsheetWorksheet } from 'google-spreadsheet';
-import WorksheetReader, { InvalidFilterError } from '../worksheet-reader';
+import WorksheetReader, { InvalidFilterError } from '../worksheet-reader.js';
 
 const createWorksheet = (title: string, index = 1) => ({
     title,
