+++ packages/core/src/reader/index.ts
@@ -1,5 +1,5 @@
-export { default as Reader } from './spreadsheet-reader';
-export type { WorksheetLinesByTitle } from './spreadsheet-reader';
+export { default as Reader } from './spreadsheet-reader.js';
+export type { WorksheetLinesByTitle } from './spreadsheet-reader.js';
 
-export { default as WorksheetReader, InvalidFilterError } from './worksheet-reader';
-export type { SheetsFilter } from './worksheet-reader';
+export { default as WorksheetReader, InvalidFilterError } from './worksheet-reader.js';
+export type { SheetsFilter } from './worksheet-reader.js';
