+++ packages/core/src/transformer/ios.ts
@@ -1,6 +1,6 @@
 import { EOL } from 'os';
-import { OutputFormat } from '../constants';
-import { Transformer } from './transformer';
+import { OutputFormat } from '../constants.js';
+import { Transformer } from './transformer.js';
 
 const iOSTransformer: Transformer = {
     outputFormat: OutputFormat.IOS,
