+++ packages/core/src/transformer/json.ts
@@ -1,6 +1,6 @@
 import { EOL } from 'os';
-import { OutputFormat } from '../constants';
-import { Transformer } from './transformer';
+import { OutputFormat } from '../constants.js';
+import { Transformer } from './transformer.js';
 
 const jsonTransformer: Transformer = {
     outputFormat: OutputFormat.JSON,
